# fastapi_celery_openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "description": "\nIES系统仿真和优化算法服务器\n\nOpenAPI描述文件(可导入Apifox): https://0.0.0.0:9870/openapi.json\nAPI文档: https://0.0.0.0:9870/docs\n",
        "version": "0.0.1"
    },
    "paths": {
        "/calculate_async": {
            "post": {
                "tags": [
                    "async"
                ],
                "summary": "异步提交能流拓扑图",
                "description": "填写数据并提交拓扑图，如果还有计算资源，提交状态为成功，返回计算ID，否则不返回计算ID，提交状态为失败",
                "operationId": "calculate_async_calculate_async_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnergyFlowGraph"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "提交状态以及模型计算ID,根据ID获取计算结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalculationAsyncSubmitResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get_calculation_state": {
            "get": {
                "tags": [
                    "async"
                ],
                "summary": "获取计算状态",
                "description": "根据计算ID获取计算状态",
                "operationId": "get_calculation_state_get_calculation_state_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Calculation Id",
                            "type": "string"
                        },
                        "name": "calculation_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Celery内置任务状态，如果是null则表示不存在该任务",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalculationStateResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get_calculation_result_async": {
            "get": {
                "tags": [
                    "async"
                ],
                "summary": "异步获取能流拓扑计算结果",
                "description": "提交计算ID，返回计算状态，如果计算完毕会一起返回数据，否则数据为空",
                "operationId": "get_calculation_result_async_get_calculation_result_async_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Calculation Id",
                            "type": "string"
                        },
                        "name": "calculation_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "计算状态和计算结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalculationAsyncResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/revoke_calculation": {
            "get": {
                "tags": [
                    "async"
                ],
                "summary": "撤销计算任务",
                "description": "提交计算ID，撤销计算",
                "operationId": "revoke_calculation_revoke_calculation_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Calculation Id",
                            "type": "string"
                        },
                        "name": "calculation_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回撤销计算状态",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RevokeResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get_calculation_ids": {
            "get": {
                "tags": [
                    "async"
                ],
                "summary": "查询任务ID",
                "description": "任务如果24小时内没有状态更新会被清出缓存，检查缓存中的所有可查询任务ID",
                "operationId": "get_calculation_ids_get_calculation_ids_get",
                "responses": {
                    "200": {
                        "description": "缓存中可查询的任务ID列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get Calculation Ids Get Calculation Ids Get",
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CalculationAsyncResult": {
                "title": "CalculationAsyncResult",
                "required": [
                    "calculation_state"
                ],
                "type": "object",
                "properties": {
                    "calculation_state": {
                        "title": "计算任务状态",
                        "enum": [
                            "PENDING",
                            "RECEIVED",
                            "STARTED",
                            "SUCCESS",
                            "FAILURE",
                            "RETRY",
                            "REVOKED"
                        ],
                        "type": "string",
                        "description": "Celery内置任务状态，如果是null则表示不存在该任务"
                    },
                    "calculation_result": {
                        "title": "计算结果",
                        "type": "object",
                        "description": "如果没有计算完或者不存在返回空，否则返回计算结果字典"
                    }
                },
                "description": "异步计算任务查询返回结果"
            },
            "CalculationAsyncSubmitResult": {
                "title": "CalculationAsyncSubmitResult",
                "required": [
                    "submit_result"
                ],
                "type": "object",
                "properties": {
                    "calculation_id": {
                        "title": "计算ID",
                        "type": "string",
                        "description": "如果成功注册计算任务，返回ID，否则为空"
                    },
                    "submit_result": {
                        "title": "提交结果",
                        "enum": [
                            "success",
                            "failed"
                        ],
                        "type": "string",
                        "description": "如果成功提交，返回\"success\"，否则返回\"failed\""
                    }
                },
                "description": "异步计算提交结果返回类"
            },
            "CalculationStateResult": {
                "title": "CalculationStateResult",
                "required": [
                    "calculation_state"
                ],
                "type": "object",
                "properties": {
                    "calculation_state": {
                        "title": "计算任务状态",
                        "enum": [
                            "PENDING",
                            "RECEIVED",
                            "STARTED",
                            "SUCCESS",
                            "FAILURE",
                            "RETRY",
                            "REVOKED"
                        ],
                        "type": "string",
                        "description": "Celery内置任务状态，如果是null则表示不存在该任务"
                    }
                },
                "description": "包含计算任务状态的数据类"
            },
            "EnergyFlowGraph": {
                "title": "EnergyFlowGraph",
                "required": [
                    "graph",
                    "nodes",
                    "adjacency"
                ],
                "type": "object",
                "properties": {
                    "graph": {
                        "title": "能流拓扑图的附加属性",
                        "type": "object",
                        "description": "仿真和优化所需的模型参数字典",
                        "examples": {
                            "建模仿真": {
                                "summary": "建模仿真所需参数",
                                "description": "建模仿真需要知道仿真步长和起始时间",
                                "value": {
                                    "模型类型": "建模仿真",
                                    "仿真步长": 60,
                                    "开始时间": "2023-3-1",
                                    "结束时间": "2024-3-1"
                                }
                            },
                            "规划设计": {
                                "summary": "规划设计所需参数",
                                "description": "规划设计不需要知道仿真步长和起始时间,会根据不同优化指标事先全部计算，不需要在此指出",
                                "value": {
                                    "模型类型": "规划设计"
                                }
                            }
                        }
                    },
                    "nodes": {
                        "title": "节点",
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "description": "由所有节点ID和属性字典组成的列表",
                        "example": [
                            {
                                "id": "a",
                                "node_type": "load"
                            },
                            {
                                "id": "b",
                                "node_type": "device"
                            },
                            {
                                "id": "c",
                                "node_type": "load"
                            },
                            {
                                "id": "d",
                                "node_type": "port",
                                "port_type": "AC"
                            },
                            {
                                "id": "e",
                                "node_type": "port",
                                "port_type": "AC"
                            },
                            {
                                "id": "f",
                                "node_type": "port",
                                "port_type": "AC"
                            }
                        ]
                    },
                    "adjacency": {
                        "title": "边",
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        },
                        "description": "由能流图中节点互相连接的边组成的列表",
                        "example": [
                            [
                                {
                                    "id": "b"
                                },
                                {
                                    "id": "d"
                                }
                            ],
                            [
                                {
                                    "id": "a"
                                },
                                {
                                    "id": "e"
                                }
                            ],
                            [
                                {
                                    "id": "c"
                                },
                                {
                                    "id": "f"
                                }
                            ],
                            [
                                {
                                    "id": "d"
                                },
                                {
                                    "id": "e"
                                }
                            ],
                            [
                                [
                                    "d",
                                    "id"
                                ],
                                {
                                    "id": "f"
                                }
                            ]
                        ]
                    }
                },
                "description": "用于仿真和优化计算的能流拓扑图，仿真和优化所需要的参数模型和变量定义会有所不同。"
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "RevokeResult": {
                "title": "RevokeResult",
                "required": [
                    "calculation_state",
                    "revoke_result"
                ],
                "type": "object",
                "properties": {
                    "calculation_state": {
                        "title": "计算任务状态",
                        "enum": [
                            "PENDING",
                            "RECEIVED",
                            "STARTED",
                            "SUCCESS",
                            "FAILURE",
                            "RETRY",
                            "REVOKED"
                        ],
                        "type": "string",
                        "description": "Celery内置任务状态，如果是null则表示不存在该任务"
                    },
                    "revoke_result": {
                        "title": "撤销结果",
                        "enum": [
                            "success",
                            "failed"
                        ],
                        "type": "string",
                        "description": "如果成功撤销任务，返回\"success\"，否则返回\"failed\""
                    }
                },
                "description": "撤销返回结果"
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}